refactor: extract OpenAIChat types to separate file to fix circular import
- Create types.ts with OpenAIChat and MultiModal interfaces
- Fix circular dependency: index.svelte.ts → hypav3.ts → index.svelte.ts
- Update ResizeBox.svelte to import getEmotion from characters module
