+++ src/lib/ChatScreens/ResizeBox.svelte
@@ -2,7 +2,7 @@
     import { CharEmotion, ViewBoxsize } from '../../ts/stores.svelte';
     import { onMount } from 'svelte';
     import TransitionImage from './TransitionImage.svelte';
-    import { getEmotion } from '../../ts/util';
+    import { getEmotion } from 'src/ts/characters';
     
     import { DBState } from 'src/ts/stores.svelte';
